Use proper connection tracking
The east-west test was failing after a fix was made to the opflex-agent.
The issue was that the test was allowing TCP sessions to pass, even though
the security groups were configured with reflexive policies. The agent was
fixed to properly implement connection tracking, which means that the TCP
sessions that start with something other than just a SYN should fail.